% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cnvs_distr_big_TF}
\alias{cnvs_distr_big_TF}
\title{Genomic distribution of CNVs (T/F)}
\usage{
cnvs_distr_big_TF(cnvs, chr_st_en, bin_size = 250000, gt)
}
\arguments{
\item{cnvs}{CNVs table}

\item{chr_st_en}{As usual}

\item{bin_size}{Size of the unit}

\item{gt}{1 for deletions, 2 fro duplications}
}
\description{
Plot the genomic distribution of deletion and duplication
genome-wide (chromosomes 1 to 22). The y axes is the rescaled
log frequency.
}
