% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_metrics_per_prob2}
\alias{get_metrics_per_prob2}
\title{Prepare performance metrics table}
\usage{
get_metrics_per_prob2(dt, unk = 3)
}
\arguments{
\item{dt}{validated CNVs table}

\item{unk}{treat unclear calls as unclear (3, excluded),
true (1), or false (2)}
}
\description{
Prepare performance metrics table
}
