% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_performance}
\alias{plot_performance}
\title{Performance metrics plot}
\usage{
plot_performance(dt)
}
\arguments{
\item{dt}{output of \code{get_metrics_per_prob2()}}
}
\description{
Performance metrics plot
}
