% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cnv_sets.R
\name{compare_cnv_sets}
\alias{compare_cnv_sets}
\title{Compare two CNV sets}
\usage{
compare_cnv_sets(cnvsA, cnvsB, gt_cn = "CN")
}
\arguments{
\item{cnvsA}{CNVs table}

\item{cnvsB}{CNVs table}

\item{gt_cn}{choose to check the exct CN or just the GT}
}
\description{
Return all pairs of overlapping segments for a given sample with the same (exact
) CN from two separate sets, e.g. two genotyping run in the same collection.
}
